consolidate API
