+++ packages/config/src/index.ts
@@ -5,4 +5,3 @@ export * from './Config.types';
 export * from './Project';
 export * from './Errors';
 export { getPrebuildConfig } from './getPrebuildConfig';
-export { getIntrospectPrebuildConfig } from './getIntrospectPrebuildConfig';
